[query] consolidate SPrimitive trait defaults + fix fromValues naming
Two cleanups in the SType primitives.

1. Move three identical overrides into the SPrimitive trait:

       castRename(t) = this
       _coerceOrCopy(...) = self-match-or-error (via assert(value.st eq this))
       settableTupleTypes() = FastSeq(ti)

   These are identical across SInt32, SInt64, SFloat32, SFloat64,
   SBoolean. The only SPrimitive implementations are those 5 case
   objects (verified by grep). Primitives have no rename-able
   virtual-type info, so defaulting castRename = this is safe (the
   footgun for SType-level pType-bearing types does not apply here).

2. Fix linguistic antipattern: fromValues(settables: ...) in SInt32,
   SInt64, SFloat64, SNDArraySlice. The trait declares 'values' as the
   parameter name; 22 of 26 implementations follow it. The parameter
   holds Values, not Settables -- Name/Type Mismatch per
   linguistic-antipatterns.md.

Each primitive file shrinks ~15-21 lines. Compiles clean on 2.12 and 2.13.
