+++ hail/hail/src/is/hail/types/physical/stypes/primitives/SBoolean.scala
@@ -1,32 +1,16 @@
 package is.hail.types.physical.stypes.primitives
 
-import is.hail.annotations.Region
 import is.hail.asm4s.{BooleanInfo, Settable, SettableBuilder, TypeInfo, Value}
 import is.hail.collection.FastSeq
 import is.hail.expr.ir.EmitCodeBuilder
 import is.hail.types.physical.{PBoolean, PType}
-import is.hail.types.physical.stypes.{SSettable, SType, SValue}
+import is.hail.types.physical.stypes.{SSettable, SValue}
 import is.hail.types.virtual.{TBoolean, Type}
 
 case object SBoolean extends SPrimitive {
   override def ti: TypeInfo[_] = BooleanInfo
 
   override lazy val virtualType: Type = TBoolean
-
-  override def castRename(t: Type): SType = this
-
-  override def _coerceOrCopy(
-    cb: EmitCodeBuilder,
-    region: Value[Region],
-    value: SValue,
-    deepCopy: Boolean,
-  ): SValue =
-    value.st match {
-      case SBoolean =>
-        value
-    }
-
-  override def settableTupleTypes(): IndexedSeq[TypeInfo[_]] = FastSeq(BooleanInfo)
 
   override def fromSettables(settables: IndexedSeq[Settable[_]]): SBooleanSettable = {
     val IndexedSeq(x: Settable[Boolean @unchecked]) = settables
