+++ hail/hail/src/is/hail/types/physical/stypes/primitives/SFloat64.scala
@@ -1,12 +1,11 @@
 package is.hail.types.physical.stypes.primitives
 
-import is.hail.annotations.Region
 import is.hail.asm4s.{DoubleInfo, Settable, SettableBuilder, TypeInfo, Value}
 import is.hail.asm4s.Code.invokeStatic1
 import is.hail.collection.FastSeq
 import is.hail.expr.ir.EmitCodeBuilder
 import is.hail.types.physical.{PFloat64, PType}
-import is.hail.types.physical.stypes.{SSettable, SType, SValue}
+import is.hail.types.physical.stypes.{SSettable, SValue}
 import is.hail.types.virtual.{TFloat64, Type}
 
 case object SFloat64 extends SPrimitive {
@@ -14,28 +13,14 @@ case object SFloat64 extends SPrimitive {
 
   override lazy val virtualType: Type = TFloat64
 
-  override def castRename(t: Type): SType = this
-
-  override def _coerceOrCopy(
-    cb: EmitCodeBuilder,
-    region: Value[Region],
-    value: SValue,
-    deepCopy: Boolean,
-  ): SValue =
-    value.st match {
-      case SFloat64 => value
-    }
-
-  override def settableTupleTypes(): IndexedSeq[TypeInfo[_]] = FastSeq(DoubleInfo)
-
   override def fromSettables(settables: IndexedSeq[Settable[_]]): SFloat64Settable = {
     val IndexedSeq(x: Settable[Double @unchecked]) = settables
     assert(x.ti == DoubleInfo)
     new SFloat64Settable(x)
   }
 
-  override def fromValues(settables: IndexedSeq[Value[_]]): SFloat64Value = {
-    val IndexedSeq(x: Value[Double @unchecked]) = settables
+  override def fromValues(values: IndexedSeq[Value[_]]): SFloat64Value = {
+    val IndexedSeq(x: Value[Double @unchecked]) = values
     assert(x.ti == DoubleInfo)
     new SFloat64Value(x)
   }
