+++ hail/hail/src/is/hail/types/physical/stypes/primitives/SInt32.scala
@@ -1,11 +1,10 @@
 package is.hail.types.physical.stypes.primitives
 
-import is.hail.annotations.Region
 import is.hail.asm4s.{IntInfo, Settable, SettableBuilder, TypeInfo, Value}
 import is.hail.collection.FastSeq
 import is.hail.expr.ir.EmitCodeBuilder
 import is.hail.types.physical.{PInt32, PType}
-import is.hail.types.physical.stypes.{SSettable, SType, SValue}
+import is.hail.types.physical.stypes.{SSettable, SValue}
 import is.hail.types.virtual.{TInt32, Type}
 
 case object SInt32 extends SPrimitive {
@@ -13,28 +12,14 @@ case object SInt32 extends SPrimitive {
 
   override lazy val virtualType: Type = TInt32
 
-  override def castRename(t: Type): SType = this
-
-  override def _coerceOrCopy(
-    cb: EmitCodeBuilder,
-    region: Value[Region],
-    value: SValue,
-    deepCopy: Boolean,
-  ): SValue =
-    value.st match {
-      case SInt32 => value
-    }
-
-  override def settableTupleTypes(): IndexedSeq[TypeInfo[_]] = FastSeq(IntInfo)
-
   override def fromSettables(settables: IndexedSeq[Settable[_]]): SInt32Settable = {
     val IndexedSeq(x: Settable[Int @unchecked]) = settables
     assert(x.ti == IntInfo)
     new SInt32Settable(x)
   }
 
-  override def fromValues(settables: IndexedSeq[Value[_]]): SInt32Value = {
-    val IndexedSeq(x: Value[Int @unchecked]) = settables
+  override def fromValues(values: IndexedSeq[Value[_]]): SInt32Value = {
+    val IndexedSeq(x: Value[Int @unchecked]) = values
     assert(x.ti == IntInfo)
     new SInt32Value(x)
   }
