+++ hail/hail/src/is/hail/types/physical/stypes/primitives/SPrimitive.scala
@@ -1,11 +1,14 @@
 package is.hail.types.physical.stypes.primitives
 
+import is.hail.annotations.Region
 import is.hail.asm4s._
 import is.hail.asm4s.implicits._
+import is.hail.collection.FastSeq
 import is.hail.expr.ir.EmitCodeBuilder
 import is.hail.io.PrefixCoder
 import is.hail.types.{RPrimitive, TypeWithRequiredness}
 import is.hail.types.physical.stypes.{SType, SValue}
+import is.hail.types.virtual.Type
 
 trait SPrimitive extends SType {
   def ti: TypeInfo[_]
@@ -15,6 +18,20 @@ trait SPrimitive extends SType {
   override def copiedType: SType = this
 
   override def containsPointers: Boolean = false
+
+  override def castRename(t: Type): SType = this
+
+  override def settableTupleTypes(): IndexedSeq[TypeInfo[_]] = FastSeq(ti)
+
+  override def _coerceOrCopy(
+    cb: EmitCodeBuilder,
+    region: Value[Region],
+    value: SValue,
+    deepCopy: Boolean,
+  ): SValue = {
+    assert(value.st eq this)
+    value
+  }
 }
 
 abstract class SPrimitiveValue extends SValue {
